apps/tg-bot/pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "vtyric",email = "misha.torno@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "telethon (>=1.40.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
